+++ srchybrid/PreferencesDlg.cpp
@@ -1,5 +1,5 @@
 //this file is part of eMule
-//Copyright (C)2002-2024 Merkur ( strEmail.Format("%s@%s", "devteam", "emule-project.net") / https://www.emule-project.net )
+//Copyright (C)2002-2026 Merkur ( strEmail.Format("%s@%s", "devteam", "emule-project.net") / https://www.emule-project.net )
 //
 //This program is free software; you can redistribute it and/or
 //modify it under the terms of the GNU General Public License
@@ -160,7 +160,7 @@ void CPreferencesDlg::Localize()
 		};
 
 		int c;
-		for (c = 0; c < _countof(uids); ++c)
+		for (c = 0; c < (int)_countof(uids); ++c)
 			LocalizeItemText(c, uids[c]);
 #if defined(_DEBUG) || defined(USE_DEBUG_DEVICE)
 		GetPageTreeControl()->SetItemText(GetPageTreeItem(c), _T("Debug"));
